refactor(settings): rename DisplaySettings module to AppearanceSettings
