+++ packages/desktop/src/renderer/pages/settings/AppearanceSettings/backgroundUtils.ts
