+++ packages/desktop/src/renderer/pages/settings/AppearanceSettings/presets.ts
