+++ packages/desktop/src/renderer/pages/settings/AppearanceSettings/presets/discourse-horizon.css
