+++ packages/desktop/src/renderer/pages/settings/AppearanceSettings/presets/glittering-input-field.css
