+++ packages/desktop/src/renderer/pages/settings/AppearanceSettings/presets/hello-kitty.css
