+++ packages/desktop/src/renderer/pages/settings/AppearanceSettings/presets/retro-windows.css
