+++ packages/desktop/src/renderer/pages/settings/AppearanceSettings/presets/retroma-nocturne-parchment.css
