+++ packages/desktop/src/renderer/pages/settings/AppearanceSettings/presets/retroma-obsidian-book-2-1-dark.css
