+++ packages/desktop/src/renderer/pages/settings/AppearanceSettings/presets/retroma-obsidian-book.css
