+++ packages/desktop/src/renderer/pages/settings/AppearanceSettings/presets/retroma-y2k.css
