+++ packages/desktop/src/renderer/pages/settings/AppearanceSettings/themeCovers.ts
